\name{createBasin}
\alias{createBasin}
\title{
creates a basin
}
\description{
instantiates an object from class of \code{createBasin}
}
\usage{
createBasin(name, simPeriod, interval)
}
\arguments{
  \item{name}{
a string: a name for the basin
}
  \item{simPeriod}{
the simulation period in seconds
}
  \item{interval}{
number of simulation time steps
}
}
\value{
a list: an object from class of \code{creatBasin}
}
\seealso{ \code{\link[RHMS]{addObjectToBasin}} }
\author{
Rezgar Arabzadeh
}
\keyword{ list }
