\name{plot.createBasin}
\alias{plot.createBasin}
\title{
plots basin layout
}
\description{
plot method for objects inherited from class of \code{createBasin}
}
\usage{
\method{plot}{createBasin}(x,...)
}
\arguments{
  \item{x}{an object from class of \code{createBasin}}
  \item{...}{other objects that can be passed to \code{plot} function}
}
\seealso{ \code{\link[RHMS]{sim}} }
\author{
Rezgar Arabzadeh
}
\keyword{ graphs }
