\name{plot.sim}
\alias{plot.sim}
\title{
plot method for an RHMS object
}
\description{
plot method for objects inherited from class of \emph{sim}
}
\usage{
\method{plot}{sim}(x,...)
}
\arguments{
  \item{x}{an object from class of \emph{sim}}
  \item{...}{other objects that can be passed to \emph{plot} function}
}
\seealso{ \code{\link[RHMS]{sim}} }
\author{
Rezgar Arabzadeh
}
\keyword{ graphs }
