\name{sim.default}
\alias{sim.default}
\title{
HMS default function
}
\description{
RHMS package default simulation function for sim class
}
\usage{
\method{sim}{default}(object)
}
\arguments{
  \item{object}{
an object from class of createBasin
}
}
\author{
Rezgar Arabzadeh
}
\seealso{ \code{\link[RHMS]{sim}} }
\keyword{ list }
