\name{summary.sim}
\alias{summary.sim}
\title{
summary method for RHMS objects
}
\description{
summary method for objects inherited from class of \code{sim}
}
\usage{
\method{summary}{sim}(object,...)

}
\arguments{
  \item{object}{an object from class of \code{sim}}
  \item{...}{other objects that can be passed to summary function}
}
\value{
a matrix: including inflow and outflow volumes and peaks rates respectively
}
\seealso{ \code{\link[RHMS]{sim}} }
\author{
Rezgar Arabzadeh
}
\keyword{ array }
