\name{Zaab}
\alias{Zaab}
\docType{data}
\title{
Zaab Subbasin datasets at Kurdistan 
}
\description{
an object inherited from class of creatBasin including datasets, reservoirs, reachs, subbasins, and junctions within a subbasin in Kurdistan known as Zaab.
}
\usage{data(Zaab)}
\source{
Iran Water Resources Management Company (2015)
}
\examples{
data(Zaab)
}
\keyword{datasets}
