\name{createBasin.default}
\alias{createBasin.default}
\title{
default function of class of \emph{createBasin}
}
\description{
instantiates an object from class of \emph{createBasin}
}
\usage{
\method{createBasin}{default}(name = "Untittled", simPeriod, interval)
}
\arguments{
  \item{name}{
a string: a name for the basin
}
  \item{simPeriod}{
the simulation period in seconds
}
  \item{interval}{
number of simulation time steps
}
}
\value{
a list: an object from class of \emph{creatBasin}
}
\seealso{ \code{\link[RHMS]{addObjectToBasin}} }
\author{
Rezgar Arabzadeh
}
\keyword{ list }
