\name{plot.sim}
\alias{plot.sim}
\title{
plot function for RHMS package
}
\description{
plot function for objects inherited from class of sim
}
\usage{
\method{plot}{sim}(x,...)
}
\arguments{
  \item{x}{an object from class of sim}
  \item{...}{other objects that can be passed to plot function}
}
\seealso{ \code{\link[RHMS]{sim}} }
\author{
Rezgar Arabzadeh
}
\keyword{ graphs }
