\name{sim.base}
\alias{sim.base}
\title{
RHMS core function
}
\description{
HMS package core simulation function for sim class
}
\usage{
\method{sim}{base}(object)
}
\arguments{
  \item{object}{
an object from class of createBasin
}
}
\author{
Rezgar Arabzadeh
}
\seealso{ \code{\link[RHMS]{sim}} }
\keyword{ list }
