\name{sim.default}
\alias{sim.default}
\title{
default function for class of \code{sim}
}
\description{
simulates an object inherited form class of \code{createBasin}
}
\usage{
\method{sim}{default}(object)
}
\arguments{
  \item{object}{
an object from class of \code{createBasin}
}
}
\author{
Rezgar Arabzadeh
}
\seealso{ \code{\link[RHMS]{sim}} }
\keyword{ list }
