\name{summary.sim}
\alias{summary.sim}
\title{
summary function for RHMS package
}
\description{
summary function for objects inherited from class of sim
}
\usage{
\method{summary}{sim}(object,...)

}
\arguments{
  \item{object}{an object from class of sim}
  \item{...}{other objects that can be passed to plot function}
}
\value{
a matrix: including inflow and outflow volumes and peaks rates respectively
}
\seealso{ \code{\link[RHMS]{sim}} }
\author{
Rezgar Arabzadeh
}
\keyword{ array }
