\name{Zaab}
\alias{Zaab}
\docType{data}
\title{
datasets for Zaab subbasin, a subbasin in Kurdistan, Iran. 
}
\description{
an object inherited from class of \code{createBasin}. including features, of a sub-basin in Kurditan known as Zaab, such as: reservoirs, reachs, subbasins, and junctions.
}
\usage{data(Zaab)}
\source{
Iran Water Resources Management Company (2015)
}
\examples{
data(Zaab)
}
\keyword{datasets}
