\name{Zaab}
\alias{Zaab}
\docType{data}
\title{
Zaab, a basin inKurdistan, datasets. 
}
\description{
an object inherited from class of \emph{createBasin}. including features, of a sub-basin in Kurditan known as Zaab, such as: reservoirs, reachs, subbasins, and junctions.
}
\usage{data(Zaab)}
\source{
Iran Water Resources Management Company (2015)
}
\examples{
data(Zaab)
}
\keyword{datasets}
